refactor(core): declutter root namespace and extract cli adapters
- Created bounded contexts `core/common`, `core/security`, and `core/telemetry`
- Moved core base, telemetry, and security models into appropriate contexts
- Extracted driving CLI adapters from `toolkit/` to dedicated `cli/` directory
- Relocated system config to `adapters/system/`
- Updated `pyproject.toml` script entry to point to new CLI main
- Executed workspace-wide search-and-replace to resolve all broken imports
- Passed Ruff formatting, linting, and Mypy static type checking
