+++ pyproject.toml
@@ -32,7 +32,7 @@ Repository = "https://github.com/CoReason-AI/coreason_manifest"
 Documentation = "https://github.com/CoReason-AI/coreason_manifest"
 
 [project.scripts]
-coreason = "coreason_manifest.cli:main"
+coreason = "coreason_manifest.cli.main:main"
 
 [dependency-groups]
 dev = [
