+++ src/coreason_manifest/core/common/base.py
