+++ src/coreason_manifest/core/oversight/mixed_initiative.py
@@ -3,8 +3,8 @@ from typing import Literal
 
 from pydantic import Field, field_validator
 
-from coreason_manifest.core.common_base import CoreasonModel
-from coreason_manifest.core.compliance import SecurityVisitor
+from coreason_manifest.core.common.base import CoreasonModel
+from coreason_manifest.core.security.compliance import SecurityVisitor
 
 
 class AllocationRule(CoreasonModel):
