+++ src/coreason_manifest/core/workflow/contracts.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from pydantic import ConfigDict, Field
 
-from coreason_manifest.core.common_base import CoreasonModel
+from coreason_manifest.core.common.base import CoreasonModel
 
 
 class AtomicSkill(CoreasonModel):
