+++ src/coreason_manifest/core/workflow/flow.py
@@ -6,12 +6,12 @@ import jsonschema  # type: ignore[import-untyped]
 from jsonschema.exceptions import SchemaError  # type: ignore[import-untyped]
 from pydantic import Field, field_validator, model_validator
 
+from coreason_manifest.core.common.base import CoreasonModel
 from coreason_manifest.core.common.semantic import SemanticRef
-from coreason_manifest.core.common_base import CoreasonModel
-from coreason_manifest.core.compliance import RemediationAction, SecurityVisitor
 from coreason_manifest.core.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.core.oversight.governance import Governance
 from coreason_manifest.core.primitives.types import MiddlewareDef, MiddlewareID, NodeID
+from coreason_manifest.core.security.compliance import RemediationAction, SecurityVisitor
 from coreason_manifest.core.state.persistence import PersistenceConfig
 from coreason_manifest.core.state.tools import AnyTool, ToolPack
 from coreason_manifest.core.workflow.nodes import (
